requirements.txt
aiohttp==3.6.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
jsonschema==3.2.0
multidict==4.7.5
pyrsistent==0.16.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
validator-collection==1.4.1
yarl==1.4.2
zipp==3.1.0
